+++ examples/i2c_command/i2c_command.ino
@@ -23,8 +23,8 @@ void setup()
 
 void loop()
 {
-  int32_t temperature;
-  int32_t pressure;
+  float temperature;
+  float pressure;
   int16_t oversampling = 7;
   int16_t ret;
   Serial.println();
